# feat: Faster and fallible query planner