--- a/apollo-router-core/src/query_planner/bridge_query_planner.rs
+++ b/apollo-router-core/src/query_planner/bridge_query_planner.rs
@@ -91,7 +91,7 @@ impl QueryPlanner for BridgeQueryPlanner {
             .planner
             .plan(query, operation)
             .await
-            .map_err(QueryPlannerError::RouterBridgeError)??;
+            .map_err(QueryPlannerError::RouterBridgeError)?;
 
         if let Some(data) = planner_result.data {
             match data {