--- a/apollo-router-core/src/query_planner/bridge_query_planner.rs
+++ b/apollo-router-core/src/query_planner/bridge_query_planner.rs
@@ -169,7 +169,7 @@ mod tests {
         let result = planner.get("".into(), None, Default::default()).await;
 
         assert_eq!(
-            "Query planning had errors: Bridge errors: UNKNOWN: Syntax Error: Unexpected <EOF>.",
+            "query planning had errors: bridge errors: UNKNOWN: Syntax Error: Unexpected <EOF>.",
             result.unwrap_err().to_string()
         );
     }