--- a/apollo-router-core/src/query_planner/mod.rs
+++ b/apollo-router-core/src/query_planner/mod.rs
@@ -1,13 +1,11 @@
 mod bridge_query_planner;
 mod caching_query_planner;
-mod router_bridge_query_planner;
 mod selection;
 use crate::prelude::graphql::*;
 pub use bridge_query_planner::*;
 pub use caching_query_planner::*;
 use fetch::OperationKind;
 use futures::prelude::*;
-pub use router_bridge_query_planner::*;
 use serde::Deserialize;
 use std::collections::HashSet;
 use tracing::Instrument;