--- a/apollo-router-core/src/services/router_service.rs
+++ b/apollo-router-core/src/services/router_service.rs
@@ -3,7 +3,7 @@ use crate::ensure_query_presence::EnsureQueryPresence;
 use crate::forbid_http_get_mutations::ForbidHttpGetMutationsLayer;
 use crate::services::execution_service::ExecutionService;
 use crate::{
-    plugin_utils, CachingQueryPlanner, DynPlugin, ExecutionRequest, ExecutionResponse,
+    plugin_utils, /*CachingQueryPlanner,*/ DynPlugin, ExecutionRequest, ExecutionResponse,
     NaiveIntrospection, Plugin, QueryCache, QueryPlannerRequest, QueryPlannerResponse,
     ResponseBody, RouterBridgeQueryPlanner, RouterRequest, RouterResponse, Schema, SubgraphRequest,
     SubgraphResponse,
@@ -243,19 +243,19 @@ impl PluggableRouterServiceBuilder {
         // the plugins in their original order.
 
         //QueryPlannerService takes an UnplannedRequest and outputs PlannedRequest
-        let plan_cache_limit = std::env::var("ROUTER_PLAN_CACHE_LIMIT")
-            .ok()
-            .and_then(|x| x.parse().ok())
-            .unwrap_or(100);
+        // let plan_cache_limit = std::env::var("ROUTER_PLAN_CACHE_LIMIT")
+        //     .ok()
+        //     .and_then(|x| x.parse().ok())
+        //     .unwrap_or(100);
 
         // QueryPlannerService takes an UnplannedRequest and outputs PlannedRequest
         let query_planner_service = ServiceBuilder::new().buffer(self.buffer).service(
             self.plugins.iter_mut().rev().fold(
-                CachingQueryPlanner::new(
-                    RouterBridgeQueryPlanner::new(self.schema.clone()),
-                    plan_cache_limit,
-                )
-                .boxed(),
+                // CachingQueryPlanner::new(
+                RouterBridgeQueryPlanner::new(self.schema.clone())
+                    // plan_cache_limit,
+                    // )
+                    .boxed(),
                 |acc, e| e.query_planning_service(acc),
             ),
         );