… CHANGELOG.md
… Cargo.lock
… apollo-router-core/Cargo.toml
… apollo-router-core/src/query_planner/router_bridge_query_planner.rs
… apollo-router-core/src/query_planner/snapshots/apollo_router_core__query_planner__bridge_query_planner__tests__empty_query_plan_should_be_a_planner_error.snap
… apollo-router-core/src/query_planner/snapshots/apollo_router_core__query_planner__bridge_query_planner__tests__plan.snap
… apollo-router/Cargo.toml
… apollo-router/src/router_factory.rs
… apollo-router/tests/snapshots/integration_tests__traced_basic_composition.snap
… apollo-router/tests/snapshots/integration_tests__traced_basic_request.snap
… examples/embedded/src/main.rs
… fuzz/Cargo.toml
… licenses.html