fixes #617

We reworked the way query plans are generated before being cached, which lead to a great performance improvement.
Moreover, the router is able to make sure the schema is valid at startup and on schema update, before you query it.