@Geal 2022-04-01
I did some benchmarks locally, still with the 5kB response used in CI.

# main, without query cache or dedup

**100 reqs, 10 clients**

Summary:
  Total:        1.6851 secs
  Slowest:      0.1848 secs
  Fastest:      0.1412 secs
  Average:      0.1647 secs
  Requests/sec: 59.3419
  
  Total data:   1192300 bytes
  Size/request: 11923 bytes

Response time histogram:
  0.141 [1]     |■■
  0.146 [0]     |
  0.150 [3]     |■■■■■■
  0.154 [10]    |■■■■■■■■■■■■■■■■■■■■
  0.159 [10]    |■■■■■■■■■■■■■■■■■■■■
  0.163 [20]    |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.167 [18]    |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.172 [18]    |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.176 [9]     |■■■■■■■■■■■■■■■■■■
  0.180 [5]     |■■■■■■■■■■
  0.185 [6]     |■■■■■■■■■■■■


Latency distribution:
  10% in 0.1528 secs
  25% in 0.1590 secs
  50% in 0.1654 secs
  75% in 0.1700 secs
  90% in 0.1770 secs
  95% in 0.1822 secs
  99% in 0.1848 secs


**1000 reqs, 100 clients**

Summary:
  Total:        21.3244 secs
  Slowest:      3.2626 secs
  Fastest:      0.8619 secs
  Average:      2.0759 secs
  Requests/sec: 46.8946
  
  Total data:   11923000 bytes
  Size/request: 11923 bytes

Response time histogram:
  0.862 [1]     |
  1.102 [8]     |■
  1.342 [10]    |■
  1.582 [37]    |■■■■
  1.822 [115]   |■■■■■■■■■■■■■
  2.062 [362]   |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  2.302 [294]   |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  2.542 [81]    |■■■■■■■■■
  2.782 [27]    |■■■
  3.022 [43]    |■■■■■
  3.263 [22]    |■■


Latency distribution:
  10% in 1.7343 secs
  25% in 1.8689 secs
  50% in 2.0413 secs
  75% in 2.2260 secs
  90% in 2.4863 secs
  95% in 2.8853 secs
  99% in 3.0928 secs

**OOM at 200 clients**

# this branch

**100 reqs, 10 clients**

Summary:
  Total:        0.4522 secs
  Slowest:      0.1347 secs
  Fastest:      0.0103 secs
  Average:      0.0423 secs
  Requests/sec: 221.1253
  
  Total data:   1192300 bytes
  Size/request: 11923 bytes

Response time histogram:
  0.010 [1]     |■
  0.023 [11]    |■■■■■■■■■■
  0.035 [45]    |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.048 [17]    |■■■■■■■■■■■■■■■
  0.060 [11]    |■■■■■■■■■■
  0.072 [3]     |■■■
  0.085 [3]     |■■■
  0.097 [1]     |■
  0.110 [3]     |■■■
  0.122 [2]     |■■
  0.135 [3]     |■■■


Latency distribution:
  10% in 0.0221 secs
  25% in 0.0288 secs
  50% in 0.0321 secs
  75% in 0.0488 secs
  90% in 0.0839 secs
  95% in 0.1148 secs
  99% in 0.1347 secs


**1000 reqs, 100 clients:**

Summary:
  Total:        3.0882 secs
  Slowest:      0.8089 secs
  Fastest:      0.0072 secs
  Average:      0.2850 secs
  Requests/sec: 323.8147
  
  Total data:   11923000 bytes
  Size/request: 11923 bytes

Response time histogram:
  0.007 [1]     |
  0.087 [27]    |■■
  0.168 [153]   |■■■■■■■■■■■
  0.248 [101]   |■■■■■■■
  0.328 [572]   |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.408 [30]    |■■
  0.488 [40]    |■■■
  0.568 [38]    |■■■
  0.649 [22]    |■■
  0.729 [11]    |■
  0.809 [5]     |


Latency distribution:
  10% in 0.1458 secs
  25% in 0.1929 secs
  50% in 0.2930 secs
  75% in 0.3050 secs
  90% in 0.4424 secs
  95% in 0.5256 secs
  99% in 0.7107 secs

**10000 reqs, 1000 clients**

Summary:
  Total:        28.3308 secs
  Slowest:      14.5817 secs
  Fastest:      0.0070 secs
  Average:      2.4824 secs
  Requests/sec: 352.9727
  
  Total data:   119230000 bytes
  Size/request: 11923 bytes

Response time histogram:
  0.007 [1]     |
  1.464 [4700]  |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  2.922 [3063]  |■■■■■■■■■■■■■■■■■■■■■■■■■■
  4.379 [1519]  |■■■■■■■■■■■■■
  5.837 [383]   |■■■
  7.294 [147]   |■
  8.752 [89]    |■
  10.209 [38]   |
  11.667 [22]   |
  13.124 [18]   |
  14.582 [20]   |
  
Latency distribution:
  10% in 1.0354 secs
  25% in 1.4032 secs
  50% in 2.1835 secs
  75% in 2.8452 secs
  90% in 4.2479 secs
  95% in 4.9576 secs
  99% in 8.5391 secs
  
  For comparison, 10000 reqs and 1000 clients on main with cache:
  
  https://github.com/apollographql/router/pull/769#issuecomment-1085637106
  
  So we're at 0.1x the RPS of main with cache